Second commit
This is my second commit and I have added all things that I wanted to add at the moment, including a raw ranking system and the ability to search for miss-matches in the genome. Also there is the options to perform more searches without having to kill the app. I'm very very happy with this commit.